feat(nm_display_28inch): add FT6336 touch support, PMU fixes and build error fixes
- board_devices.yaml: add lcd_touch device (FT6336/ft5x06, I2C 0x70, swap_xy+mirror_y for landscape)
- setup_device.c: add lcd_touch_factory_entry_t using esp_lcd_touch_new_i2c_ft5x06
- setup_device.c: fix AXP2101 PMU power-path config; raise VBUS input current limit
  to 1500 mA (REG 0x16) so the board runs stably from USB-C without a battery;
  configure charger control (REG 0x18) and battery detection (REG 0x68)
- sdkconfig.defaults.board: enable APP_CLAW_LUA_MODULE_LCD_TOUCH for Lua touch bindings
- http_reuse: guard esp_http_client_is_persistent_connection stub for IDF < 5.5
- claw_event_router: fix -Werror=int-in-bool-context in ESP_LOG_LEVEL ternary
- esp_painter: fix -Werror=format= with explicit (uint32_t) cast
- display_dirty: add missing stddef.h for NULL
- lua_driver_touch: wrap touch_sens API in SOC_TOUCH_SENSOR_VERSION==3 guard
  (ESP32-S3 is V2 and not supported by the new driver; compile stubs instead)
- lua_driver_touch CMakeLists: remove manual include dirs (no longer needed)

+++ components/common/esp_painter/esp_painter.c
@@ -257,7 +257,7 @@ static esp_err_t esp_painter_draw_pixel(esp_painter_handle_t handle, uint8_t *bu
 #endif
 
     if (using_ppa) {
-        ESP_LOGD(TAG, "Drawing pixel at (%d,%d) with color 0x%04" PRIx32, x, y, color_value);
+        ESP_LOGD(TAG, "Drawing pixel at (%d,%d) with color 0x%04" PRIx32, x, y, (uint32_t)color_value);
 
         uint32_t offset = (y * painter->text_canvas.width + x) * 2;
         if (offset + 2 > buffer_size) {
